+++ builder_pkgs/mockito/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 5.5.1
 
-* Allow `source_gen: ">=3.0.0 <5.0.0"`.
+* Require `analyzer: ^8.1.0`.
+* Allow `source_gen: '>=3.0.0 <5.0.0'`.
+* Allow `build: '>=3.0.0 <5.0.0'`.
 
 ## 5.5.0